+++ src/impala/rle-encoding.h
@@ -231,7 +231,7 @@ inline bool RleDecoder::Get(T* val) {
   if (UNLIKELY(literal_count_ == 0 && repeat_count_ == 0)) {
     // Read the next run's indicator int, it could be a literal or repeated run
     // The int is encoded as a vlq-encoded value.
-    uint32_t indicator_value = 0;
+    uint64_t indicator_value = 0;
     bool result = bit_reader_.GetVlqInt(&indicator_value);
     if (!result) return false;
